requirements.txt
numpy
Pillow
torch>=1.2
tabulate
tqdm
torchvision
biopython
transformers